Phaser "body.removeShape" triggers error of undefined
This is just a fast fix but fixed my issue of a shape which gets temporarely added to a body. I guess it triggers this issue (si.position of undefined) when removed but still in loop. Anyhow, there probably is a much better fix but I don't have time to dig deeper. Hope you can have a look at this issue!

Thanks